revert move
